UCS-120 : removed AttributeMonitor component since its task can be accomplished by RM
