+++ UCSCore/src/main/java/it/cnr/iit/ucs/core/UCSCoreServiceBuilder.java
@@ -93,7 +93,6 @@ public class UCSCoreServiceBuilder {
         ucsCore.contextHandler.setPdp( ucsCore.pdp );
         ucsCore.contextHandler.setObligationManager( ucsCore.obligationManager );
         ucsCore.contextHandler.setPIPs( new ArrayList<PIPCHInterface>( ucsCore.pipList ) );
-        ucsCore.contextHandler.setMonitoringThread( true );
         ucsCore.requestManager.setContextHandler( ucsCore.contextHandler );
         ucsCore.requestManager.setPEPMap( ucsCore.pepMap );
         for( PIPBase pip : ucsCore.pipList ) {
